// src/enum/resources.enum.ts
enum Resource{
    AUTH = "auth", // 1
    USER = "user", // 2
    OTP = "otp", // 3
    KEUANGAN = "keuangan", // 4
    KURIKULUM = "kurikulum", // 5
    LAPORAN = "laporan", // 6
    JURNAL = "jurnal", // 7
    PENILAIAN = "penilaian", // 8
    CUTI = "cuti", // 9
    JADWAL = "jadwal", // 10
    KHS = "khs", // 11
    KRS = "krs", // 12
    NILAI = "nilai", // 13
    PEMBAYARAN = "pembayaran", // 14
    TUGAS = "tugas", // 15
    WISUDA = "wisuda", // 16
    NOTIFIKASI = "notikasi", // 17
    UTIL = "util", // 18

    NONE = "none", // 19
}
export default  Resource