// src/enum/action.enum.ts
enum Action{
    INDEX = "index", // 1
    CREATE = "create", // 2
    SHOW = "show", // 3
    EDIT = "edit", // 4
    DELETE = "delete", // 5
        
    RESET = "reset", // 6
    FREEZE = "freeze", // 7
    
    GENERATE = "generate", // 8
    VERIFY = "verify", // 9
   
    LOGIN = "login", // 10
    LOGOUT = "logout", // 11

    PING = "ping", // 12
    NONE = "none", // 13
}
export default Action